Refactors Docker configuration and CI workflow
Streamlines the Docker setup by copying the virtual environment directly and running the bot with Python directly, eliminating the need for sourcing the virtual environment in the CMD.

This simplifies the Dockerfile and improves efficiency. Also, adds debug information to the CI workflow for better tag management and environment awareness during builds.

Commit for test CICD